please.sh: handle uncommitted `pkgver` changes in `maybe_force_pkgrel()`
When `pkgver` was changed, there is no need to edit `pkgrel`. Therefore,
let's just skip the logic to auto-increment `pkgrel` if `pkgver` was
changed already (as indicated by `git blame` returning the pseudo rev
`0000000`), there is no need to auto-increment the `pkgrel`.

While at it, ignore uncommitted modifications of the `pkgrel`: if the
`pkgrel` was re-set to `1` during a regular `please.sh upgrade
<package>` run, we still want to auto-increment it from the previously
committed number.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>