render-release-notes: add anchors for all versions
It is pretty convenient to be able to slap on a '#v2.14.1' and jump
immediately to the corresponding release notes.

For good measure, throw in `#latest` (this needs a slight adjustment in
the CSS because the header text is now wrapped within an `<a>` tag).

While at it, also support `#known-issues` and `#licenses`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>