render-release-notes: show Git for Windows' logo
It was "reported" on Twitter (which is not a bug tracker, last time we
checked) that the release notes' color scheme might make it look like a
Blue Screen of Death. Hopefully the Git for Windows logo will break up
that impression. If not, well, at least we show the logo now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>