nuget: force a valid package version
We want to extend the Azure Pipeline that we use to build official Git
versions, so that it can also build snapshot versions.

This not only will consolidate Pipelines, but it will also build NuGet
packages for our snapshot versions "for free".

To prepare for that, the script we use to build those NuGet packages
needs to learn to handle the "prerelease-*" style versions gracefully,
though, so let's teach it precisely that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>