git update-git-for-windows: make silent installation very silent
The command successfully identifies and downloads the update however when it
triggers the update it causes a license screen to display which requires
user interaction. Therefore the update can't be fully automated.

This is due to the installer only using the /SILENT option on the new
installer. Let's add also the /VERYSILENT option so that the installer
works as expected.

Signed-off-by: Leo D'Arcy <leo.darcy@poweronplatforms.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>