Release Notes: highlight the link to the upstream release notes
It's not immediately obvious that the "Comes with Git v2.x.y" in the
ReleaseNotes.md file showing the various upstream Git versions is a
link and, more imporatntly, that the link is to the upstream release
notes, which users ought read.

Tell the reader that this link is to to the upstream release notes.

A subseqent patch will update older entries in the ReleaseNotes.md file.

Signed-off-by: Philip Oakley <philipoakley@iee.email>
---
the extra text could be outside the link text