git update-git-for-windows: never restart silently
Git for Windows' own update command has a mode where you can check
whether a new Git for Windows version is available and automatically
install it silently if there is one.

As of e119d03cf (git update-git-for-windows: make silent installation
very silent, 2020-03-20), this does not even ask whether it is allowed
to reboot (InnoSetup suggests to reboot if it replaces any files that
had been in use during the install).

Let's prevent that from happening: automatic reboots without any warning
are just not nice.

This fixes https://github.com/git-for-windows/git/issues/2604

Helped-by: Matthias Aßhauer <mha1993@live.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>