Release Notes: notice snapshot releases
The release notes' top line shows the most recently released official
version. When snapshot versions are released, that version is identical
to the one mentioned in the first "Changes since" section.

The code assumed that they were not identical, though, and erroneously
assigned the same version to the first two "Changes since" sections.

Notice the duplication of the version and assign the ID 'snapshot'
instead. This ID can be used as anchor, e.g.
file:///path/to/ReleaseNotes.html#snapshot

Signed-off-by: Philip Oakley <philipoakley@iee.email>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>