ReleaseNotes: highlight the Upstream release notes links
The "* Comes with Git v2.x.y" in the link does not do justice to the
reality that this is _the_ link to the upstream's release notes for
the key bug fixes, improvements and backward compatibility issues in
this release.

Explicitly state that this link is for the Upstream release notes.

Editing done programmatically to match the previous commit that updated
the maintainer script for any further mentions of newe Git versions using
$ sed -r 's|(^\* Comes with \[Git v[0-9.]*)\]|\1 \(Upstream release notes\)\]|'
  <ReleaseNotes.md >tmp.md
, then inspecting and renaming in place.

Signed-off-by: Philip Oakley <philipoakley@iee.email>