ReleaseNotes: highlight the upstream Product release notes links
The "* Comes with Git v2.x.y" in the link does not do justice to the
reality that this is _the_ link to the upstream's release notes for
the key bug fixes, improvements and backward compatibility issues in
this release.

Explicitly state that this link is for the Product release notes.

Editing is performed here to avoid affecting the maintainer scripts.

Signed-off-by: Philip Oakley <philipoakley@iee.email>