ReleaseNotes: show version number/id adjacent the collapsible arrow
The release notes show the "Changes Since", without indicating the
current version number for which the changes apply.

Add the change's current version after the collapsible arrow cursor,
so as to appear floating in the top left of the level 2 header box.

While at it revert the arrow size, and hence displayed version number,
to its default size.

Signed-off-by: Philip Oakley <philipoakley@iee.email>