please.sh upgrade msys2-runtime: fix bundle creation
Just after a merging-rebase, we need to be able to create the Git
bundle. However, we're on a "detached HEAD" at that time, but we really
want to update the `main` branch. So let's do that.

Technically, this might have ramifications beyond the msys2-runtime
repository, but even in other cases, `main` would be the appropriate
branch to update.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>