commit-msys2.sh: in the oneline, don't count `git-extra`
The `git-extra` package is very special in that it not only brings a
couple things like `git update-git-for-windows`, but it also has a
post-install script that does _tons_ of fixups on top of MSYS2 that Git
for Windows could not otherwise be customized.

For the benefit of fixing up things in that post-install script, we
_always_ reinstall it at the end of the full daily update, and the
`desc` file will therefore always have at least an updated "installed
on" date.

As a consequence, we already ignore the `git-extra` package when
summarizing the package updates in the commit message's body.

Note: technically, we overdo things by ignoring even real updates of the
`git-extra` package. In the interest of not overcomplicating the script,
we leave this for later, if it ever is deemed necessary.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>