pacman-helper: work around changes in Pacman v5.2.2
In Pacman's new version, `repo-add` only uses `$GPGKEY` quoted. However,
Git for Windows' pacman-helper relied on the fact that it was _not_
quoted, so that e.g. `--batch` could be passed via that environment
variable.

Granted, this is a hacky usage, but there seems not to be another way to
pass these options to `repo-add`'s GPG calls.

To reinstate the desired behavior, add another hack on top: we now
copy-rewrite `repo-add` (which is a shell script) to a version that uses
`$GPGKEY` unquoted again.

This is needed, quickly, to unblock us from including a new GCM Core
version in Git for Windows' SDK.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>