please.sh sync: also ensure that Christoph's key is known
The original idea was of course to add David's and Christoph's new keys
to the keyring _long_ before the keys are actually used (because
otherwise, how would you get them? The package to update the keyring
would be signed with the key that you would not yet have).

Unfortunately, this did not happen in the i686 version of the packages,
only in the x86_64 ones.

Since Git for Windows wants to continue to provide 32-bit versions, we
have to work around this.

This bug held up Git for Windows v2.28.0 by a bit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>