credential-helper-selector: use the correct format specifier for Unicode
There are several instances where we want to format a string with
Unicode parameters (i.e. encoded in UTF-16). For those, we actually need
to use `%S` (with an upper-case `S`).

For some reason, this seemed to work with a lower-case `%s` before, but
it probably shouldn't have worked.

As a consequence, the choices were not persisted correctly.

This fixes https://github.com/git-for-windows/git/issues/2776.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>