minimal-sdk: reduce `PATH` to the absolute bare minimum
Some of Git's tests have prerequisites depending on what is found on the
`PATH`. For example, the GPG-related tests first try to execute `gpg`
and if it fails, they are skipped. Since GPG-related operations seem to
be not so common, that is an acceptable set of tests to skip, in
particular given how much time it would take otherwise (if the `PATH`
contains Subversion or GPG, that will change what tests are run).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>