installer: add wordpad as an editor choice
To propperly work as an editor for git we need to teach the existing
wrapper script some new tricks. Thankfully we don't need to reinvent the
wheel and can pick up the EOL handling from the notepad wrapper. We
don't need to convert files to CRLF before passing files to wordpad,
but when we get them back we need to convert them back to LF. While
we're at it let's also pick up the commit message formating.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>