installer: fix rendering of <RED>/<A> in paragraphs
Funnily enough, `GetTextWidth()` produces a non-zero width for Line Feed
charaters.

When determining the horizontal offset while rendering the colored text,
we do accumulate in the `RowPrefix` variable all uncolored text, except
Carriage Return characters (which re-set the `RowPrefix` variable to the
empty string).

Crucially, we also include Line Feed characters, Which means that our
rendering is screwed up when using the `<RED>` or `<A>` tag in
everything but the first line of each paragraph.

Let's work around that issue simply by skipping the Line Feed characters
(but only after a Carriage Return, every other instance would not make
sense anyway).

Happily, this does not affect any existing label. But we are about to
introduce a new label that would be affected, so let's fix the bug Right
Now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>