make-file-list.sh: exclude recode-sr-latin
For some reason, `mingw-w64-gettext` contains the awfully specific tool
`recode-sr-latin` (which converts Serbian to Latin). We do not need it
in Git for Windows, and nobody could have used it, anyway, as it links
to `libgettextlib-0-19-8-1.dll`, which we already exclude from the Git
for Windows installer.

So let's just exclude it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>