make-file-list.sh: exclude a SASL helper
The `sasldblistusers2` tool wants to list users in sasldb, however, we
do not ship with db, therefore that tool would not have worked, if any
user had ever tried to run it for whatever reason.

This means that that executable is dead weight, and we're better off not
even including it in Git for Windows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>