make-file-list.sh: exclude the HPACK tools from MinGit
The HPACK tools `inflatehd` and `deflatehd` are provided by the nghttp2
project, but they are not actually necessary to let libcurl talk to
HTTP/2 servers.

In MinGit, they would have failed to start, anyways, as they depend on
`libstdc++-6.dll`, which we exclude from MinGit.

So let's not bundle them.

These tools were the only users of `libjemalloc.dll`, which we therefore
exclude at the same time.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>