Add a tool to detect missing DLLs from make-file-list.sh's output
We recently built a Git for Windows snapshot with a broken
`git-remote-https`: it was missing a `.dll` file because
`mingw-w64-curl`'s package metadata failed to indicate that it depends
on `libzstd.dll`.

In the meantime, we fixed that problem. However, it took manual testing
to detect the breakage, and we would like to have an easier time finding
this type of bug in the future.

So let's add a script that runs through the output of
`make-file-list.sh` to see whether any `.dll` file is missing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>