check-for-missing-dlls.sh: optionally parse `ldd` output
This patch is the result of a (failed) experiment: in the 32-bit SDK, it
seems that `objdump -p` isn't able to parse the `.exe` files that were
processed by `cv2pdb` (the symptom is "file format not recognized") so
the next logical step was to try with `ldd` instead (which unfortunately
failed because e.g. `ldd usr/lib/perl5/core_perl/auto/Cwd/Cwd.dll` hangs
while waiting for `ldh.exe` to do something or other).

Let's keep the code, though, it might come in handy at some stage.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>