check-for-missing-dlls: also check for unneeded DLLs
At the same time when we verify that all `.exe` files bundled in Git for
Windows are accompanied by their `.dll` dependencies, we can also ensure
that we do not include `.dll` files that are not actually needed.

Note that we need to filter the generated list manually, as .NET
dependencies are not handled correctly by `objdump`, and the `.dll`
files associated with Perl/Tcl modules are obviously needed to keep
those modules working even if no `.exe` file depends on them directly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>