installer: _really_ upgrade GCM(W) users to GCM Core
The `IsDowngrade()` function is really a slippery slope: it takes two
arguments, but if even one of them has four numbers in it, the
parameters are ignored and the installed Git version is used for the
second parameter, and the current display version for the first one.

This is _very_ dangerous, as it lured this developer into using this
function with a different order of arguments.

Let's use the `IsUpgrade()` function instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>