installer: do show auto-upgraded GCM Core as 'unseen page'
When a user upgrades Git for Windows, having previously chosen Git
Credential Manager for Windows (also known as "GCM(W)"), we now upgrade
them to Git Credential Manager Core (also known as "GCM Core").

Arguably, this is "a previously unseen option", and therefore the page
should be shown when that checkbox to show only such options is checked.

We already tried that, but added the wrong thing to the set of pages
with unseen options (we need to add the page ID).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>