installer: when spawning `cmd`, disable `AutoRun`
As per the documentation, usually `AutoRun` steps specified via the
registry (e.g. `Software\Microsoft\Command Processor\AutoRun`) are run
when calling `cmd.exe`, unless the `/D` option is used:
https://docs.microsoft.com/en-us/windows-server/administration/windows-commands/cmd

This can lead to problems when we use `cmd.exe` for the benefit of
redirecting the output to a file: under certain circumstances it can
easily produce an error "The system cannot find the path specified."

Let's use the `/D` to address this problem.

This addresses https://github.com/git-for-windows/git/issues/2384.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>