please.sh upgrade: it is okay not to generate artifacts
The primary users of this sub-command are Git for Windows' Azure
Pipelines that are used to build the Pacman packages and upload them to
Git for Windows' Pacman repositories.

Naturally, those Pipelines want to generate the artifacts, so that the
commits can be pushed when everything succeeded (and only then).

But it is not a mistake to run this sub-command manually, without
generating those artifacts. Yet the script would fail in that case. Fix
this.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>