please.sh upgrade: handle for Perl modules after Perl upgrades
Whenever a major Perl version is released, the Perl modules have to be
re-built because the name of the Perl `.dll` changes (its major version
is embedded in the file name).

Therefore, a major Perl version is reason enough to increase the
`pkgrel` of the Perl modules we build in Git for Windows.

This patch automates this, so that it will be enough in the future to
trigger new builds for the Perl modules after a new `perl` was built.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>