installer: record the default choice for the Default Branch option
If the user wants to simply follow what Git wants to use as the default
branch name in freshly-initialized repositories, we used to record the
empty string.

Except we didn't.

Because InnoSetup's `SetPreviousData()` function special-cases the empty
string to mean "delete that key from the previous data instead of
recording a value" (even if the documentation does not spell that out:
https://jrsoftware.org/ishelp/index.php?topic=isxfunc_setpreviousdata).

Let's just use a single space character instead of the empty string
(which does not specify a valid branch name) to indicate that the user
did not want to override Git's default.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>