installer: refuse illegal default branch names
When the user wants to override the default branch name, it is all too
easy to make a mistake and slip e.g. a trailing space in. Let's help
users by preventing such mistakes.

Note: technically the correct way would _not_ be to reimplement
`check_ref_format()`, but at the stage when we want to let the user
specify the branch name, we do not necessarily have a working `git.exe`
to call. That's why the logic was reimplemented (without regular
expressions! Yay Pascal!).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>