helper-selector: fix quoting of trailing backslashes
In Git for Windows' 6d8684161ee9 (mingw: fix quoting of arguments,
2019-09-13), we fixed a bug where it would quote command-line arguments
with trailing backslashes incorrectly when they needed to be quoted.

While the helper-selector only quotes paths to files (which, by
definition, cannot end in a backslash), we still should fix the
`quote()` function, just in case that we will want to use it on
parameters other than file paths.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>