please.sh upgrade msys2-runtime: fix logic to determine up-to-date-ness
The `please.sh` script has a rather simple logic to determine whether a
new package should be built or not: comparing the timestamps of the
latest commits of the `msys2-runtime` repository and of the
`msys2-runtime/` subdirectory of the `MSYS2-packages` repository.

However, what we need to look for is a commit in `MSYS2-packages` that
changed either `pkgrel` or `pkgver` in `msys2-runtime/PKGBUILD`.

Let's do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>