please.sh upgrade: be a bit more forgiving in the pkgrel check
Since 77c30eedb (please.sh upgrade: make sure not to decrement pkgrel,
2020-10-14), we try to ensure that the `pkgrel` is not decremented. To
do so, the `-pkgrel=<N>` and the `+pkgrel=<N>` lines are used, deleting
the `pkgrel=` part and evaluating the remainder as an arithmetic
expression.

However, we were a bit overzealous, even taking a difference of 0 as a
sign of pkgrel being decremented.

Never mind that this cannot happen in practice (if both `<N>` are
identical, there are no -/+ lines, it will become a context line), an
empty string in `test 0 -ge <string>` is treated as if it was 0.

So let's be a bit more forgiving about this: a pkgrel "change" of
plus/minus zero is oh-kay.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>