installer/portable: unbreak update-ca-trust
Running the script `/usr/bin/update-cu-trust` requires a p11-kit `.dll`
that we excluded recently from the installer and from the Portable Git.

The reason for that: this `.dll` is loaded at runtime, it is not a
compile-time dependency. So let's include it specifically (or, more
prescisely, exclude the other p11-kit `.dll`, the one that we still do
not want).

This fixes https://github.com/git-for-windows/git/issues/2874

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>