minimal-sdk: restrict PATH
The main purpose of the `minimal-sdk` artifact is to serve in automated
Azure Pipelines and GitHub workflows. Let's make sure that we are not
picking up any unintended executables along the way.

One such unintended executable would be Git for Windows' gpg.exe: Git
for Windows is installed in the build agents of Azure Pipelines and
GitHub Actions, and its `/usr/bin/` is sadly in the `PATH` and will most
likely remain there indefinitely, for details, see
https://github.com/actions/virtual-environments/issues/1525.

However, this can lead to `gpg.exe` being picked up by Git's test suite,
and unfortunately the GPG agent tends to hang in the VMs in which the
tests are run.

So why does this not happen all the time? Most likely due to mismatches
of the MSYS2 runtime, which let `gpg.exe` fail to load (and which Git's
test suite takes as sign that "we looked all around but no GPG could be
found"). This is a fragile situation, of course: at any point, outside
of Git's (or Git for Windows') control, `gpg.exe` can start working
again.

By specifically overriding the `PATH` to only include the `bin/`
directories of the minimal SDK, plus Windows' system directories, we can
avoid this from ever happening again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>