sdk init git: skip hard-linking the built-ins by default
When initializing the Git directory, we initialize a `config.mak` file
with good default options for contributors. The `SKIP_DASHED_BUILT_INS`
knob is a new such option: it saves quite a bit of time when building
Git incrementally, and does not have any adverse effects (unless a
contributor uses a super old script that still uses the "dashed" form to
call Git commands that has been deprecated for over a decade).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>