please.sh coverity: work around Coverity issue with GCC v10.x
It is becoming exceedingly clear that Synopsys is not interested in
fixing that bug any time soon where GCC v10.2.0's version number is
mis-parsed as "0.2.0". Mistaking this for a super-ancient GCC, Coverity
completely gives up and reports no issues whatsoever.

Let's work around that by forcing the cached results to represent the
GCC version correctly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>