git update-git-for-windows: escape installer arguments
When installer is being executed MSYS2 subsystem tries to convert unix paths to Windows paths.

It seems it applies to installer arguments as well (because they start with slash).
Thus `/SILENT` is converted to `C:/Program Files/Git/SILENT` (or something alike).

This effectively bypasses installer arguments.

The solution is to escape slashes in installer arguments.