credential-helper-selector: really use the correct format specifier
In b64ba6a6d (credential-helper-selector: use the correct format
specifier for Unicode, 2020-08-11), we changed all those `%s` to `%S` to
make sure that wide-character string parameters were handled correctly.

While this does appear to work correctly, it is a bit fragile because
`%S` is fragile: there are two standards defining its behavior, and they
disagree.

The Microsoft convention is that `%S` in a wide-character format string
means that the argument is a multi-byte string:
https://docs.microsoft.com/en-us/cpp/c-runtime-library/format-specification-syntax-printf-and-wprintf-functions

This is the opposite interpretation of POSIX, which defines `%S` to
indicate a wide-character string argument, whether specified in a
multi-byte format string or a wide-character format string:
https://pubs.opengroup.org/onlinepubs/9699919799/functions/fwprintf.html

However, there is one safe recourse: if you want to print wide-character
strings via `printf()` or `wprintf()`, just use the `%ls` format
specifier. That works with POSIX semantics as well as with Microsoft
semantics.

So let's do exactly that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>