credential-helper-selector: quote path correctly
In general, we cannot guarantee that the path of the chosen credential
helper has no special characters in it. Therefore, we have to quote it.

This can be verified by putting a credential manager into a directory
whose path contains a space (such as `C:\Program Files\...`), adding
that directory to the `PATH`, and then calling the
credential-helper-selector e.g. with the argument `version` (which at
least Git Credential Manager Core understands).

While the executable will still be spawned correctly, it will then fail
to parse its command line correctly.

The symptom looks like this:

	Unrecognized command 'Files'.

	usage: git-credential-manager-core <command>

	  Available commands:
	    erase
	    get
	    store

	    configure [--system]
	    unconfigure [--system]

	    --version, version
	    --help, -h, -?

Let's quote the absolute path correctly when copying it into the command
line.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>