git update-git-for-windows: make installation just silent again
Currently the installer is started with `/VERYSILENT` switch which
causes no indication of installation to be displayed.

This may be confusing to the user because he doesn't know if the
installer is still in the progress or if it finished already.

Use of the switch was introduced in e119d03cf because of:

> The command successfully identifies and downloads the update however when it
> triggers the update it causes a license screen to display which requires
> user interaction. Therefore the update can't be fully automated.

Though it seems that running installer with just `/SILENT` doesn't
require any user interaction. Only installation progress is displayed.