Update aliases.sh
Recent windows versions have the improved console which reports as xterm-256color (and also Windows Terminal reports the same), and they definitely do not need winpty to translate anything. This PR fixes to check so that it applies solely to mintty. Also, given that mintty is not shipped with GfW (anymore?) maybe remove this file completely?