installer: allow overriding the default branch name
When generating a new installer, this lets the user specify what default
branch to use by default, via the `DEFAULT_BRANCH_NAME` constant.

The installer will then set that default in the system config. This is
necessary because `git init` will warn loudly about an unset
`init.defaultBranch` as of Git v2.30.0-rc1 (and we're already asking the
user in the installer, so we don't want to pester them again when they
run `git init`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>