please.sh: accommodate for changes in GitHub API's output
Granted, it is a bit fragile what we do here, parsing JSON via `sed`. A
better alternative would be to rely on a tool like `jq`, but it might
not be available (and we don't want to add 3.4MB to the Git for Windows
payload just to parse JSON).

So let's make the `sed` invocation a _little_ bit more forgiving. We
still expect the `"id"` keys of the individual releases to be indented
by exactly four spaces, and we still expect the "id" keyword to come
before the "tag_name" keyword. That's a risk we'll accept for the time
being.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>