ever-green: fix bug in the fixup!/squash! check
Before finishing the rebase, we want to make sure that no fixup!/squash!
commits are left. To do that, we enumerate all such commits in the
post-rebase range.

However, we still have a replace ref in place that was intended to avoid
skipping branches from being merged that were part of the
pre-merging-rebase commit history (Git would determine that those
branches are already part of the history, but the start of the merging
rebase would have reverted all of their changes).

And if a branch had been merged that was based on a _previous_ Git for
Windows version, one whose commit history contains fixup!/squash!
commits, this replace ref would interfere with our careful check.

This very issue prevented `ever-green.sh` from finishing rebases
successfully after fde3e95e30be (Merge pull request #2966 from
dennisameling/fix-double-exe, 2021-01-05) in Git for Windows, which
merged a branch based on v2.29.2.windows.3 into the main branch that had
already advanced to v2.30.0.windows.1.

Therefore, let's make sure that this replace ref is deleted before we
perform that fixup!/squash! check.

We do this via a `sed` script that inserts the `git replace --delete`
command just before the fixup!/squash! check, if one is found, otherwise
appends that command at the end of the todo list.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>