git-extra: allow to extend i686 support to ARM64
Windows/ARM64 can run i686 executables via a built-in emulator. Since it
is currently not possible to build much of anything in MSYS2 for
Windows/ARM64, let's sort of support Windows/ARM64 via starting with an
i686 version of Git for Windows' SDK, then augmenting it by Git
artifacts built via Visual Studio and installed into `/arm64/*`.

To allow for that, we now handle the special value `MSYSTEM=ARM64` to
set up the environment accordingly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>