sdk create-desktop-icon: use the appropriate API
Given OneDrive with the Desktop Redirection feature, and other supported
methods to NOT have your Desktop in %USERPROFILE%\Desktop, we should use
a more sophisticated method to get the path to the Desktop in order to
put a shortcut there. Since we are using Bash we need to step out and
invoke a vbs script to get the SpecialFolder called Desktop, and use
that as a path.

Previously, 89fe672a9 (sdk create-desktop-icon: use correct path to
Desktop, 2020-06-25) queried the registry, which probably works, too,
but it is an implementation detail and it is better to use the correct
Win32 API.

Original-patch-by: Kristof Mattei <github@kristofmattei.be>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>