pacman-helper.sh: auto-sanitize databases during `quick_add`
Git for Windows' Pacman repository is actually _never_ modified
manually. It is always updated via a dedicated Azure Pipeline that uses
`quick_add` to upgrade packages with versions that were just built.

Use the same Pipeline to sanitize the databases, and for good measure,
keep doing that (`sanitize_db` amounts to a no-op if nothing needs to be
sanitized).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>