mingit: support building from core Git branches again
Core (or "upstream") Git still has the system config in what amounts to
`/mingw64/etc/gitconfig` (or `/mingw32/etc/gitconfig`). We do have
fall-back code for that, but it was incorrect: when trying to make the
parent directory of said system config, it cut off one component from
the wrong end of the absolute path.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>