installer: improve error message when files are in use in silent mode
In silent mode, the wizard form is not shown. But unless running in
"very silent" mode, error messages _are_ shown.

This poses a problem when Git for Windows' files are in use, as the
error message refers to the "Processes" page (which is not visible in
silent mode).

Let's enhance the error message specifically for the silent mode.

This addresses https://github.com/git-for-windows/git/issues/3118

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>