icon: add more resolutions to accommodate UI scaling on high DPI devices
While Windows should be able to scale down most desired icon resolutions
from the ones we already include, it sometimes struggles to do so. Thanks
to PNG compression which is supported on Windows Vista and above, these
additional resolutions don't even take up more space than the previous
version of the Git for Windows icon.

Let's include the standard icon sizes (16x16, 32x32 and 48x48) scaled to
support scaling factors 100% to 200% at 25% intervals and 200% to 500% at
50% intervals. This adds  the following resolutions:

* 16x16 (small icons at 100%)
* 20x20 (small icons at 125%)
* 28x28 (small icons at 175%)
* 40x40 (small icons at 250%, medium icons at 125%)
* 56x56 (small icons at 350%, medium icons at 175%)
* 60x60 (large icons at 125%)
* 72x72 (small icons at 450%, large icons at 150%)
* 80x80 (small icons at 500%, medium icons at 250%)
* 84x84 (large icons at 175%)
* 96x96 (medium icons at 300%, large icons at 200%)
* 112x112 (medium icons at 350%)
* 144x144 (medium icons at 450%, large icons at 300%)
* 160x160 (medium icons at 500%)
* 168x168 (large icons at 350%)
* 192x192 (large icons at 400%)
* 216x216 (large icons at 450%)
* 240x240 (large icons at 500%)

This addresses git-for-windows/git#3184

Signed-off-by: Matthias Aßhauer <mha1993@live.de>