git update-git-for-windows: optionally update from a fork
For the upcoming ARM64 support, as well as for the Microsoft fork of Git
for Windows, it would be helpful if there was a way to update from the
GitHub Releases of the respective fork instead of Git for Windows' own
fork.

To facilitate that, let's introduce support for a
`git-update-git-for-windows.config` file next to the script: if it
exists, treat it as a Git config file and read the `update.fromFork`
value, and then continue with that value inserted into the URL
https://github.com/<fork>/releases.

Example: to ask the updater to look for new versions at
https://github.com/dennisameling/git/releases, run the following command
as administrator:

	git config \
	   -f "$(git --exec-path)/git-update-git-for-windows.config" \
	   update.fromFork dennisameling/git

This also overrides the label "Git for Windows" with "Git"; The
`update.gitLabel` setting in the same `.config` file can be used to
over-override this.

If the `.config` file does not contain any `update.fromFork`, the
file is ignored altogether.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>