git update-git-for-windows: download ARM64 Git for Windows if applicable
If the /arm64/bin folder is present, we assume that the user has the
ARM64 version of Git for Windows installed. When a new version is
available, we'll then automatically download the ARM64 version. This is
similar to the logic in git-wrapper.c in mingw-w64-git (technically, the
Git wrapper _also_ tests whether it is running an ARM64 executable, but
we cannot do that in a shell script).

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>