installer: include "Windows Terminal profile" in new options
When the checkbox "Only show new options" is selected, we want to skip
pages with options that the user has seen during previous installations
already.

With the "Windows Terminal profile" option, it is a bit complicated
because it is presented as a _component_.

This puts in a little extra hack to show the Components page when
upgrading from versions prior to v2.32.0 (which is where that Windows
Terminal profile was introduced).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>