installer: offer `scalar.exe` as a component, if found
We are in the process of porting Scalar from th .NET version to C. To
enable us to ship this in an experimental microsoft/git release, let's
detect that `scalar.exe` _is_ in the set of included files. And if it
is, offer it as a component (which is installed by default).

Technically, the scalar files are included in the payload, therefore it
is not _really_ a component. Just like Git LFS, if the component is
unchecked, we therefore remove those just-installed files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>