installer: refactor out a function to run a command silently
We call `schtasks` via `cmd.exe` for the sole purpose of being able to
redirect the `stdout` and `stderr` just in case that anything goes
wrong, so that we can show that output (but only in case of a problem).

This paradigm is very useful, and is already used in two places: when
scheduling the auto-updater, and when removing that scheduled task.

Let's refactor this into its own function.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>