installer: automatically reconfigure Scalar enlistments
The idea is that new Scalar versions might need to set additional config
options, or change existing ones. Calling `scalar reconfigure --all`
lets Scalar do this in all currently registered Scalar enlistments.

Since the Scalar enlistments are registered on a per-user level, we need
to start the `reconfigure` command as the original user. Technically,
this skips Scalar enlistments if there are other users on the same
machine who also have registered Scalar enlistments. In practice, this
is most likely a moot issue because there are only few multi-user
development setups. In multi-user setups, the other users will have to
run `scalar reconfigure --all` manually after an upgrade, and only if
the upgraded Scalar needs enlistments to be reconfigured.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>